refactor: Switch out useContext for useState
Overkill
